Allow configurable logging
Load in an external logging configuration
rather than write to a fixed file location. This
defaults to writing to the console which plays
nicer with systemd.

Remove the logging of the last received
inverter and summary requests from the
Enphase Envoy. These are logged to the main
log as before.

Log the version at start up at INFO level.

Tests:
 - Run with the default config and verify that
    the console output matches the previous
    log file output
 - Lower the debug level to ERROR and verify
    the console is essentially quiet in operation